Created MeterTest
